refactor: 패키지 구조를 도메인 중심으로 개선
