+++ src/test/java/com/example/solidconnection/country/fixture/CountryFixture.java
@@ -1,6 +1,7 @@
-package com.example.solidconnection.support.fixture;
+package com.example.solidconnection.country.fixture;
 
 import com.example.solidconnection.entity.Country;
+import com.example.solidconnection.region.fixture.RegionFixture;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestComponent;
