+++ src/test/java/com/example/solidconnection/country/repository/CountryRepositoryForTest.java
@@ -1,12 +1,10 @@
-package com.example.solidconnection.support.fixture;
+package com.example.solidconnection.country.repository;
 
 import com.example.solidconnection.entity.Country;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface CountryRepositoryForTest extends JpaRepository<Country, Long> {
 
     Optional<Country> findByCode(String code);
