+++ src/test/java/com/example/solidconnection/region/fixture/RegionFixtureBuilder.java
@@ -1,6 +1,7 @@
-package com.example.solidconnection.support.fixture;
+package com.example.solidconnection.region.fixture;
 
 import com.example.solidconnection.entity.Region;
+import com.example.solidconnection.region.repository.RegionRepositoryForTest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.test.context.TestComponent;
