+++ src/test/java/com/example/solidconnection/university/fixture/LanguageRequirementFixture.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.support.fixture;
+package com.example.solidconnection.university.fixture;
 
 import com.example.solidconnection.type.LanguageTestType;
 import com.example.solidconnection.university.domain.LanguageRequirement;
