+++ src/test/java/com/example/solidconnection/university/fixture/UniversityInfoForApplyFixture.java
@@ -1,4 +1,4 @@
-package com.example.solidconnection.support.fixture;
+package com.example.solidconnection.university.fixture;
 
 import com.example.solidconnection.university.domain.UniversityInfoForApply;
 import lombok.RequiredArgsConstructor;
